CBSOData3.jl

Documentation for CBSOData3.jl

CBSOData3.ODataTable — Method

ODataTable(table; typed, columns, filters, base, api, catalog)

Get a Tables object for the dataset defined by table. Optional parameters are:

typed::Bool, do we want a TypedDataSet (numbers as ints or floats) or an UntypedDataSet (everything string)

columns::Vector{String}, exact columnnames to select, empty for all columns

filters::String, OData3 specification for row filter

base::String, basename of the OData3 server

api::String, path part of the OData3 server for regular data

catalog::String, path part of the OData3 server for catalog information

Examples

tbl = CBSOData3.ODataTable("82811NED", columns=["Perioden", "Onderzoekspopulatie_1", "Innovatoren_2"], filter="Perioden eq '2010X000'");

CBSOData3.get_meta — Method

get_meta(table; base, api)

Get the metadata for a given table. The result is a dict with the keys "TableInfos", "DataProperties" and all the classifications.

Optional parameters are:

base::String, basename of the OData3 server

api::String, path part of the OData3 server for regular data

CBSOData3.get_tables — Function

get_tables(base, catalog)

Get all the tables in the catalog as a list of dicts. The Identifier can be used to get the actual data. If no parameters are given, the data is retrieved from the (CBS)[https://www.cbs.nl] OData3 portal.

The optional parameters are:

base::String, basename of the OData3 server

catalog::String, path part of the OData3 server for catalog information